refactor: filter_section.dart 분리
